refactoring

+++ StudentsMVC/Controllers/LessonsController.cs
@@ -32,7 +32,7 @@ namespace Students.MVC.Controllers
         public async Task<IActionResult> Index(string sortRecords, string searchString, int skip, int take, EnumPageActions action, EnumSearchParametersLesson serachParameter)
         {
             ViewData["searchString"] = searchString;
-            ViewData["serachParameter"] = serachParameter;
+            ViewData["serachParameter"] = (int)serachParameter;
             return View(_mapper.Map<IEnumerable<LessonViewModel>>((await _lessonService.DisplayingIndex(action, searchString, (EnumSearchParameters)(int)serachParameter, take, skip))));
         }
         #endregion
