+++ StudentsMVC/Controllers/ManagerController.cs
@@ -33,7 +33,7 @@ namespace Students.MVC.Controllers
         public async Task<IActionResult> Index(string sortRecords, string searchString, int skip, int take, EnumPageActions action, EnumSearchParametersManager serachParameter)
         {
             ViewData["searchString"] = searchString;
-            ViewData["serachParameter"] = serachParameter;
+            ViewData["serachParameter"] = (int)serachParameter;
             return View(_mapper.Map<IEnumerable<ManagerViewModel>>((await _managerService.DisplayingIndex(action, searchString, (EnumSearchParameters)(int)serachParameter, take, skip))));
         }
         #endregion
